requirements.txt
astroid==2.9.3
attrs==23.1.0
aws-cdk-lib==2.70.0
aws-cdk.aws-amplify-alpha
aws-cdk.aws-cognito-identitypool-alpha
aws-cdk.aws_apigatewayv2_alpha 
aws-cdk.aws_apigatewayv2_integrations_alpha
aws_cdk.aws_apigatewayv2_authorizers_alpha
cdk_nag
constructs>=10.0.0,<11.0.0
boto3==1.27.0
botocore==1.30.0
cattrs==1.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
docopt==0.6.2
idna==3.3
importlib-metadata==4.11.2
iniconfig==1.1.1
isort==5.10.1
jmespath==0.10.0
jsii<2.0.0,>=1.77.0
lazy-object-proxy==1.7.1
mccabe==0.6.1
more-itertools==8.12.0
nodeenv==1.6.0
numpy==1.21.5
packaging==21.3
pandas==1.3.5
platformdirs==2.5.1
pluggy==1.0.0
publication==0.0.3
py==1.11.0
pylint==2.12.2
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
sagemaker==2.169.0
s3transfer==0.6.1
six==1.16.0
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.2
typing_extensions==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
wrapt==1.13.3
yarg==0.1.9
zipp==3.7.0